Squashed commit of the following:
commit 5d7ac7b522bf68289d5abd78c7a79f23df188bb3
Author: Chuanzhou Su <2523901380@qq.com>
Date:   Wed May 6 17:40:53 2026 +0800

    打印数据

commit d3353087be472e8a24a3190a9f3668d531ef3aae
Author: Chuanzhou Su <2523901380@qq.com>
Date:   Tue Apr 28 23:08:38 2026 +0800

    视频回放

commit 3741d8ecd44f46466a2e45eb94012ac41591fc75
Author: Chuanzhou Su <2523901380@qq.com>
Date:   Mon Apr 27 22:49:01 2026 +0800

    修改一点摄像头扫描逻辑，目前只扫描9段

commit c2e8b1ef81507ddf88c6111f0c8c69c9cefde51a
Author: Chuanzhou Su <2523901380@qq.com>
Date:   Sun Apr 26 21:06:48 2026 +0800

    修改存储位置

commit f82930e6cf235ac32c28c66bed60064dba97feed
Author: Chuanzhou Su <2523901380@qq.com>
Date:   Sun Apr 26 20:59:53 2026 +0800

    设置页面修改

commit e100fa3f1556545a8f95c65d6b9a41f4d66eb84e
Author: Chuanzhou Su <2523901380@qq.com>
Date:   Sun Apr 26 19:12:48 2026 +0800

    md修改

commit 7012105e19510b5b90211d4b9cdfb24288bd23a1
Author: Chuanzhou Su <2523901380@qq.com>
Date:   Sun Apr 26 18:56:25 2026 +0800

    适配智能分析ai

commit dc65d2a602e98a3713ac3db77d354ee39d8f39a9
Author: Chuanzhou Su <2523901380@qq.com>
Date:   Sun Apr 26 18:38:57 2026 +0800

    ai智能分析报表

+++ app/models/db_models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Double, Text, JSON, func
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Double, Text, JSON, func
 from app.core.database import Base
 
 
@@ -77,7 +77,7 @@ class FaceHistory(Base):
     dominant_emotion = Column(String(30))
     confidence = Column(Double)
     timestamp = Column(DateTime, nullable=False, server_default=func.now())
-    frame_timestamp = Column(Integer, default=0)  # 毫秒级时间戳，用于视频对齐
+    frame_timestamp = Column(BigInteger, default=0)  # 毫秒级时间戳，用于视频对齐
     valence = Column(Double, default=0)
     arousal = Column(Double, default=0)
     pleasure = Column(Double, default=0)
